Make sure output generated in the print-object method always goes to the right stream.
There was an implicit reliance on being able to pass nil as a stream and
have the string output simply be returned; in the case of print this is
not true. Also a duration with non-zero nsecs should not be considered
as an empty duration. Made stream argument to human-readable-duration
required.